+++ docs/changelog.md
@@ -35,6 +35,7 @@ nav_order: 2
 
 - Use `slots` in addresses, Telegram, DPTBinary, DPTArray, TPCI, APCI, DPTComplexData
 - Convert Telegram and APCI to dataclasses. `Telegram` is not hashable anymore.
+- Use `asyncio.gather` where applicable (mostly device telegram process pipeline)
 
 # 2.12.2 Fix thread leak 2024-03-05